requirements.txt
pdfplumber>=0.10.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
requests>=2.31.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-cov>=4.0.0

# Image intelligence (OCR, captioning, embeddings)
pytesseract>=0.3.10
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
open_clip_torch>=2.20.0
